Currency Overview, Svg Graph 데이터 만들던 중

+++ src/main/java/HwangJiHun/poeitemvalues/model/ninja/SparkLine.java
@@ -8,6 +8,6 @@ import java.util.List;
 @Data
 public class SparkLine {
 
-    private List<Float> data = new ArrayList<>(7);
-    private Float totalChange;
+    private List<Double> data = new ArrayList<>(7);
+    private Double totalChange;
 }
